Remove ow specific code in packet shipper/sorter + buffer reader
