+++ BattleNetwork/overworld/bnPacketShipper.h
@@ -6,36 +6,42 @@
 #include <chrono>
 #include <vector>
 
-namespace Overworld
+enum class Reliability : char
 {
-  class PacketShipper
+  Unreliable = 0,
+  UnreliableSequenced,
+  Reliable,
+  ReliableSequenced,
+  ReliableOrdered,
+};
+
+class PacketShipper
+{
+private:
+  struct BackedUpPacket
   {
-  private:
-    struct BackedUpPacket
-    {
-      uint64_t id;
-      std::chrono::time_point<std::chrono::steady_clock> creationTime;
-      Poco::Buffer<char> data;
-    };
-
-    Poco::Net::SocketAddress socketAddress;
-    uint64_t nextUnreliableSequenced;
-    uint64_t nextReliable;
-    uint64_t nextReliableOrdered;
-    std::vector<BackedUpPacket> backedUpReliable;
-    std::vector<BackedUpPacket> backedUpReliableOrdered;
-    bool failed;
-
-    void sendSafe(Poco::Net::DatagramSocket& socket, const Poco::Buffer<char>& data);
-    void acknowledgedReliable(uint64_t id);
-    void acknowledgedReliableOrdered(uint64_t id);
-
-  public:
-    PacketShipper(const Poco::Net::SocketAddress& socketAddress);
-
-    bool HasFailed();
-    void Send(Poco::Net::DatagramSocket& socket, Reliability Reliability, const Poco::Buffer<char>& body);
-    void ResendBackedUpPackets(Poco::Net::DatagramSocket& socket);
-    void Acknowledged(Reliability reliability, uint64_t id);
+    uint64_t id;
+    std::chrono::time_point<std::chrono::steady_clock> creationTime;
+    Poco::Buffer<char> data;
   };
-} // namespace Overworld
+
+  Poco::Net::SocketAddress socketAddress;
+  uint64_t nextUnreliableSequenced;
+  uint64_t nextReliable;
+  uint64_t nextReliableOrdered;
+  std::vector<BackedUpPacket> backedUpReliable;
+  std::vector<BackedUpPacket> backedUpReliableOrdered;
+  bool failed;
+
+  void sendSafe(Poco::Net::DatagramSocket& socket, const Poco::Buffer<char>& data);
+  void acknowledgedReliable(uint64_t id);
+  void acknowledgedReliableOrdered(uint64_t id);
+
+public:
+  PacketShipper(const Poco::Net::SocketAddress& socketAddress);
+
+  bool HasFailed();
+  void Send(Poco::Net::DatagramSocket& socket, Reliability Reliability, const Poco::Buffer<char>& body);
+  void ResendBackedUpPackets(Poco::Net::DatagramSocket& socket);
+  void Acknowledged(Reliability reliability, uint64_t id);
+};
